pyproject.toml
[tool.poetry]
name = "artems-diploma"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.37"
python-telegram-bot = "^13.12"
simplegmail = "^4.0.4"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
